Refactored HadoopStoreJobRunner to use jopt and improved '-output' parameter behavior.
It was agreed that it's a bug to just dump the built store into an existing directory. Now you either must pass a non-existent directory or use the '-force-overwrite' flag to make sure it's wiped before running job.

+++ contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/HadoopStoreBuilder.java
@@ -131,8 +131,14 @@ public class HadoopStoreBuilder {
         FileOutputFormat.setOutputPath(conf, tempDir);
 
         try {
+
+            FileSystem fs = outputDir.getFileSystem(conf);
+            if(fs.exists(outputDir)) {
+                throw new IOException("Final output directory already exists.");
+            }
+
             // delete output dir if it already exists
-            FileSystem fs = tempDir.getFileSystem(conf);
+            fs = tempDir.getFileSystem(conf);
             fs.delete(tempDir, true);
 
             long size = sizeOfPath(fs, inputPath);
